Create The Number of Visible Pairs
The Number of Visible Pairs

Given a doubly circular linked list with non-negative integers. Each integer represents the height of a mountain. For example, {1->2->3->4->5}, {2->3->4->5->1} or {3->4->5->1->2} are the same in the circular linked list.

If two mountains A and B can be mutually visible, they need to meet the following criteria.
   1. A and B must not be the same mountain.
   2. When A and B are adjacent, they definitely can see each other, In the example above, (1,2), (2,3), (3,4), (4,5) and (1,5) are adjacent. This, we have five visible pairs.
   3. When A and B are not adjacent and let hmin denote as their minimum height (either A's or B's), they are mutually visible if the height of any mountain between A and B in either clockwise or anti-clockwise is not higher than hmin. In the example above, (2,5) and (3,5) are two pairs which meet this criteria.

Another example is [5->2->2->4->3]. In this example, (2,5), (2,2), (2,4), (3,4) and (3,5) are adjacent. They are visible. Moreover, (4,5), (2,5) (third node and first node), (2,4) (second node and fourth node). Thus, they are eight mutually visible pairs.

Write a function to find the number of visible pairs in a given doubly circular linked list. It is noted that you only need to find the number.

int numMountainPairs(CDblLinkedList CDLL);

Sample input
3
1 2 3 4 5 a
5 2 2 4 3 a
5 4 4 4 3 2 1 5 1 a

Sample output
Current List: 5 4 3 2 1
7 Pairs.
Current List: 3 4 2 2 5
8 Pairs.
Current List: 1 5 1 2 3 4 4 4 5
18 Pairs.